restructure flask app
