+++ flaskApp/app/__init__.py
@@ -14,7 +14,7 @@ def create_app():
     app = Flask(__name__)
     
     # Load configuration
-    app.config.from_object('app.config.Config')
+    app.config.from_object('config.Config')
 
     # Initialize components
     db.init_app(app)
@@ -22,8 +22,14 @@ def create_app():
 
     Swagger(app)
 
-     # Import and register the Blueprint
-    from .routes import bp  # Import the Blueprint
-    app.register_blueprint(bp)
+    from app.routes.cheapest_route import cheapest_route_bp
+    from app.routes.cheapest_vehicle import cheapest_vehicle_bp
+    from app.routes.gate_by_id import gate_by_id_bp
+    from app.routes.gates import gates_bp
+
+    app.register_blueprint(gates_bp)
+    app.register_blueprint(gate_by_id_bp)
+    app.register_blueprint(cheapest_route_bp)
+    app.register_blueprint(cheapest_vehicle_bp)
 
     return app
