+++ flaskApp/app/models.py
@@ -1,12 +1,12 @@
 from . import db
 
-
-# Define a Sample Model
+# Model to hold info on individual gate
 class Gate(db.Model):
     __tablename__ = 'gate'
     id = db.Column(db.String, primary_key=True)
     name = db.Column(db.String, nullable=False)
 
+# Model to hold info on connections between gates
 class GateConnection(db.Model):
     __tablename__ = 'gate_connections'
     id = db.Column(db.Integer, primary_key=True)
